Revert "fix: 弹幕拉取改用 protobuf seg.so，修复 XML 空壳导致合并源 0 条"
This reverts commit 6294d88a901e214b383287cdb2363b0ec22523a0.
