+++ registry/lib/components/video/danmaku/merger/api/bilibili.ts
@@ -1,13 +1,26 @@
-import { bilibiliApi, getJsonWithCredentials } from '@/core/ajax'
+import { bilibiliApi, getJsonWithCredentials, monkey } from '@/core/ajax'
 import type { PageItem, SearchResult, ViewResult } from './types'
-import { fetchDanmakuItems, getDanmakuXml } from './danmaku-fetch'
-
-export { isPakkuActive } from './pakku'
-export { fetchDanmakuItems, getDanmakuXml }
 
 const SEARCH_URL = 'https://api.bilibili.com/x/web-interface/search/type'
 const VIEW_URL = 'https://api.bilibili.com/x/web-interface/view'
 const PAGE_LIST_URL = 'https://api.bilibili.com/x/player/pagelist'
+const DANMAKU_XML_URL = 'https://comment.bilibili.com'
+
+/** 是否检测到 pakku.js 注入（会劫持页面 XHR/fetch 的弹幕请求） */
+export const isPakkuActive = (): boolean => {
+  try {
+    if (document.querySelector('.__pakku_injected')) {
+      return true
+    }
+    // pakku 在 content script 里给 XHR 打上 pakku_open
+    if ((XMLHttpRequest.prototype as { pakku_open?: unknown }).pakku_open) {
+      return true
+    }
+  } catch {
+    // 环境受限时忽略
+  }
+  return false
+}
 
 /** 解析 view 查询参数：支持 BV 与 av 号（与 runtime 直输 BV/av 一致） */
 const buildViewUrl = (id: string): string => {
@@ -46,4 +59,22 @@ export const getView = async (bvid: string): Promise<ViewResult> => {
 export const getPageList = async (bvid: string): Promise<PageItem[]> => {
   const url = `${PAGE_LIST_URL}?bvid=${encodeURIComponent(bvid)}`
   return bilibiliApi<PageItem[]>(getJsonWithCredentials(url), '获取分P列表失败')
+}
+
+/** 获取弹幕 XML（comment.bilibili.com，与 runtime getDanmaku 一致） */
+export const getDanmakuXml = async (cid: number | string): Promise<string> => {
+  // 必须走 GM_xmlhttpRequest：pakku.js 会劫持页面 XHR/fetch 的 .xml 弹幕请求，
+  // 对「非当前播放 cid」的 comment.bilibili.com/{cid}.xml 可能一直不返回，导致合并/恢复卡死。
+  const responseText = await monkey<string>({
+    url: `${DANMAKU_XML_URL}/${cid}.xml`,
+    method: 'GET',
+    responseType: 'text',
+    anonymous: false,
+  })
+
+  if (responseText == null || responseText === '') {
+    throw new Error('弹幕接口返回为空')
+  }
+
+  return responseText
 }
