+++ registry/lib/components/video/danmaku/merger/api/types.ts
@@ -35,7 +35,6 @@ export interface ViewPageItem {
 /** view 接口视频详情 */
 export interface ViewResult {
   bvid: string
-  aid?: number
   title: string
   pic: string
   owner?: { name: string }
