+++ registry/lib/components/video/danmaku/merger/ui/quick-merge-host.ts
@@ -49,10 +49,7 @@ export interface QuickMergeHostDeps {
       owner?: { name: string }
       pages?: Array<{ cid: number; part: string; page: number }>
     }>
-    getDanmaku: (
-      cid: number | string,
-      aid?: number | string,
-    ) => Promise<Array<{ time: number; text: string }>>
+    getDanmaku: (cid: number | string) => Promise<string>
   }
   parseDanmaku: (xml: string) => Array<{ time: number; text: string }>
   injectDanmaku: (
@@ -132,14 +129,8 @@ export const createQuickMergeHost = (deps: QuickMergeHostDeps) => {
       }
 
       const p1 = data.pages[0]
-      const list = await deps.api.getDanmaku(
-        p1.cid,
-        (data as { aid?: number | string }).aid,
-      )
-      if (!list.length) {
-        mergerToast(`并入失败：${data.title}（弹幕为空）`, 'error')
-        return
-      }
+      const xml = await deps.api.getDanmaku(p1.cid)
+      const list = deps.parseDanmaku(xml)
 
       const result = await deps.injectDanmaku(
         list,
@@ -147,7 +138,6 @@ export const createQuickMergeHost = (deps: QuickMergeHostDeps) => {
           id: `${bvid}_${p1.cid}`,
           cid: p1.cid,
           bvid,
-          aid: (data as { aid?: number | string }).aid,
           title: data.title + (data.pages.length > 1 ? ' P1' : ''),
           pic: normalizeHttpsUrl(data.pic),
           author: data.owner?.name || '',
