+++ registry/lib/components/video/danmaku/merger/ui/vue-host.ts
@@ -61,10 +61,7 @@ export interface MergerVueHostDeps {
       owner?: { name: string }
       pages?: Array<{ cid: number; part: string; page: number; duration?: number }>
     }>
-    getDanmaku: (
-      cid: number | string,
-      aid?: number | string,
-    ) => Promise<Array<{ time: number; text: string; type?: number; color?: number; size?: number; date?: number; uid?: string; dmid?: string }>>
+    getDanmaku: (cid: number | string) => Promise<string>
     getPageList: (bvid: string) => Promise<Array<{ cid: number; duration?: number }>>
   }
   parseDanmaku: (xml: string) => Array<{ time: number; text: string }>
@@ -551,7 +548,8 @@ export const createMergerVueHost = (deps: MergerVueHostDeps): MergerVueHostCtrl 
     await Promise.all(
       pages.map(async page => {
         try {
-          const list = await deps.api.getDanmaku(page.cid)
+          const xml = await deps.api.getDanmaku(page.cid)
+          const list = deps.parseDanmaku(xml)
           page.danmakuCount = list.length
           page.danmakuError = false
         } catch {
@@ -744,7 +742,6 @@ export const createMergerVueHost = (deps: MergerVueHostDeps): MergerVueHostCtrl 
                 pic: normalizeHttpsUrl(data.pic),
                 author: data.owner?.name || '',
                 groupTitle: data.title,
-                aid: (data as { aid?: number | string }).aid,
               })
             } else {
               throw new Error('没有分P数据')
@@ -756,10 +753,8 @@ export const createMergerVueHost = (deps: MergerVueHostDeps): MergerVueHostCtrl 
           }
           resolvedTitle = title
           resolvedSourceId = workingTask.bvid ? `${workingTask.bvid}_${cid}` : String(cid)
-          const list = await deps.api.getDanmaku(cid, workingTask.aid)
-          if (!list.length) {
-            throw new Error('弹幕为空')
-          }
+          const xml = await deps.api.getDanmaku(cid)
+          const list = deps.parseDanmaku(xml)
           const injectIndex = success + fail + 1
           const result = await deps.injectDanmaku(
             list,
